🔧 Rename(#13): rename folder names to common-components
