+++ src/app/test/page.tsx
@@ -2,11 +2,11 @@
 'use client'
 
 import { useState } from "react";
-import PurpleInput from "@component/components/input/PurPleInput";
-import GrayInput from '@component/components/input/GrayInput';
-import PurpleTextarea from "@component/components/textarea/Textarea";
-import TabComponent from "@component/components/tab/TabComponent";
-import Dropdown from "@component/components/dropdown/Dropdown";
+import PurpleInput from "@component/components/common-components/input/PurPleInput";
+import GrayInput from '@component/components/common-components/input/GrayInput';
+import PurpleTextarea from "@component/components/common-components/textarea/Textarea";
+import TabComponent from "@component/components/common-components/tab/TabComponent";
+import Dropdown from "@component/components/common-components/button/dropdown/Dropdown";
 import LandingTop from "@component/components/landing/carousel/LandingTop";
 
 const Test = () => {
