+++ src/components/common-components/input/PurPleInput.types.ts
