+++ src/components/common-components/tab/Tab.types.ts
